Add razor pages

+++ Models/Reservation/ReservationChambre.cs
@@ -3,6 +3,6 @@
     public class ReservationChambre : Reservation
     {
         public Produit.Chambre Chambre { get; set; }
-        public decimal Prix => Chambre.PrixJournalier * CalculerDureeReservation();
+        public int Prix => Chambre.PrixJournalier * CalculerDureeReservation();
     }
 }
