+++ Models/Reservation/ReservationVoiture.cs
@@ -5,6 +5,6 @@ namespace GestionReservations.Models.Reservation
     public class ReservationVoiture : Reservation
     {
         public Produit.Voiture Voiture { get; set; }
-        public decimal Prix => Voiture.PrixJournalier * CalculerDureeReservation();
+        public int Prix => Voiture.PrixJournalier * CalculerDureeReservation();
     }
 }
